LoggingPanel: apply text filtering to recorded text and log messages
This requiers to actually store the text and log messages that are
received by LoggingPanel, in correct order and with meta data.
(LogRecorder is used for this purpose.)

Whenever the user changes the text to filter for, it is neccessary to
reprocess and redisplay the recorded text and messages. This might
need to much time to be processed in the swing thread, therefore the
new class ItemTextPane is introduced. It can display an expanding
list of items, and uses and worker thread when the list is replaced.